feat: align GantryDetailPage with boost interactions
Replace upvotes with boost/pin on the detail page. Extract
CardBoostersSection into a shared BoostersSection component reused
by both RoadmapCard and GantryDetailPage. Admins see an inline
pinners list (avatars, names, notes) when pinCount > 0.
